Move funcion has_glitch from rnog_utilities (removed) to RNO_G/channelGlitchDetector.py
